Working with dates and lambda expressions
Task 0. The number of the day of the week is displayed (1 is Sunday).
Task 1. From the given date is the date of the next Tuesday.
Task 2. Implemented a generalized functional interface. As a result, either the string is displayed in reverse order, or the factorial of the number (depending on the user's choice).